checkasm: fix type of the "counter" argument of some functions

+++ tests/checkasm/sdidec.c
@@ -38,7 +38,7 @@ static void randomize_buffers(uint8_t *src0, uint8_t *src1)
 void checkasm_check_sdidec(void)
 {
     struct {
-        void (*uyvy)(const uint8_t *src, uint16_t *dst, int64_t pixels);
+        void (*uyvy)(const uint8_t *src, uint16_t *dst, uintptr_t pixels);
     } s = {
         .uyvy = upipe_sdi_to_uyvy_c,
     };
@@ -59,7 +59,7 @@ void checkasm_check_sdidec(void)
         uint8_t  src1[NUM_SAMPLES * 10 / 8];
         DECLARE_ALIGNED(32, uint16_t, dst0)[NUM_SAMPLES];
         DECLARE_ALIGNED(32, uint16_t, dst1)[NUM_SAMPLES];
-        declare_func(void, const uint8_t *src, uint16_t *dst, int64_t pixels);
+        declare_func(void, const uint8_t *src, uint16_t *dst, uintptr_t pixels);
 
         randomize_buffers(src0, src1);
         call_ref(src0, dst0, NUM_SAMPLES / 2);
